requirements.txt
ccxt==3.0.84
openai==1.4.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pandas
streamlit
plotly